data: Move selected hunk filtering
Hunk navigation still selects and recalculates hunks, while selected-
change loading now sits in its own data module.

It still owns batch attribution filtering for cached line-level hunks,
which keeps ownership masking and consumed replacement masks coupled to
navigation.

This commit continues narrowing hunk_tracking by moving cached selected-
hunk filtering into data.selected_change.hunk_filtering and updating
show/include importers.

Subsequent commits will move the filtering tests and update architecture
guards for the new owner.
